+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ## 0.0.2-dev
 
-- Fixed embeded the excel template
+- Fixed embedding the Excel template
 - Fixed ARB format
 
 ## 0.0.3-dev
@@ -25,6 +25,6 @@
 - Added command line option to define the *lead locale* used as the source locale in translation excels created.
 - Migrated to Flutter 3.27.x and newer versions of used Excel Plugin.
 - Added support for importing and exporting selected languages only (including and not including the lead locale resources).
-- Added option for filtering resources depending on meta tag (e.g. "x-reviewed": false) and option to merge existing
+- Added option for filtering resources depending on review markers (e.g. "x-reviewed": false) and option to merge existing
   ARB file with translated contents from Excel.
 - Removed useless -n command line option.